requirements.txt
gradio>=4.0.0
groq>=0.4.0
elevenlabs>=0.2.0
gtts>=2.3.0
speechrecognition>=3.10.0
pydub>=0.25.1
python-dotenv>=1.0.0

